requirements.txt
openai
langgraph
pydantic-ai
gradio
pypdf
python-dotenv
